Added more undocumented functions
